+++ frontend/src/pages/channels/index.ts
@@ -3,12 +3,12 @@ import { NavItemEntry } from '../types'
 import { channelsPageRoute } from '../routes'
 
 import ChannelsPage from './channels-page'
+import { isAnyChannelOn } from './channels-actions'
 
 export const channelsPageNavItem: NavItemEntry = {
   route: channelsPageRoute,
   icon: iconFader,
   label: 'Channels',
   page: ChannelsPage,
-  highlighted: apiState =>
-    apiState.channels?.some(value => value !== 0) ?? false,
+  highlighted: isAnyChannelOn,
 }