+++ frontend/src/pages/chases/chases-actions.tsx
@@ -6,6 +6,7 @@ import { isAnyOn, isAllOn } from '../../util/state'
 import { HotkeyContext } from '../../hooks/hotkey'
 import { memoInProduction } from '../../util/development'
 import { apiState } from '../../api/api-state'
+import { ApiState } from '../../api/worker/processing'
 
 function setOnForAllChases(on: boolean) {
   Object.entries(apiState.liveChases).forEach(([id, liveChase]) => {
@@ -15,8 +16,12 @@ function setOnForAllChases(on: boolean) {
   })
 }
 
-const turnAllOn = () => setOnForAllChases(true)
-const turnAllOff = () => setOnForAllChases(false)
+export function isAnyChaseOn(apiState: ApiState) {
+  return isAnyOn(apiState.liveChases)
+}
+
+export const turnAllChasesOn = () => setOnForAllChases(true)
+export const turnAllChasesOff = () => setOnForAllChases(false)
 
 /**
  * Corner actions for the chases page:
@@ -27,7 +32,7 @@ export const ChasesActions = memoInProduction(() => {
   const { allOn, allOff } = useApiStateSelector(
     apiState => ({
       allOn: isAllOn(apiState.liveChases),
-      allOff: !isAnyOn(apiState.liveChases),
+      allOff: !isAnyChaseOn(apiState),
     }),
     { event: 'liveChases' }
   )
@@ -36,7 +41,7 @@ export const ChasesActions = memoInProduction(() => {
     <HotkeyContext.Provider value={true}>
       <Button
         icon={iconLight}
-        onClick={turnAllOn}
+        onClick={turnAllChasesOn}
         disabled={allOn}
         title="All on"
         hotkey="o"
@@ -45,7 +50,7 @@ export const ChasesActions = memoInProduction(() => {
       </Button>
       <Button
         icon={iconLightOff}
-        onClick={turnAllOff}
+        onClick={turnAllChasesOff}
         disabled={allOff}
         title="All off"
         hotkey="p"