+++ frontend/src/pages/chases/index.ts
@@ -1,14 +1,14 @@
 import { iconChase } from '../../ui/icons'
-import { isAnyOn } from '../../util/state'
 import { NavItemEntry } from '../types'
 import { chasesPageRoute } from '../routes'
 
 import ChasesPage from './chases-page'
+import { isAnyChaseOn } from './chases-actions'
 
 export const chasesPageNavItem: NavItemEntry = {
   route: chasesPageRoute,
   icon: iconChase,
   label: 'Chases',
   page: ChasesPage,
-  highlighted: apiState => isAnyOn(apiState.liveChases ?? {}),
+  highlighted: isAnyChaseOn,
 }