+++ frontend/src/pages/fixture-groups/index.ts
@@ -1,14 +1,14 @@
 import { iconGroup } from '../../ui/icons'
-import { isAnyOn } from '../../util/state'
 import { NavItemEntry } from '../types'
 import { fixtureGroupsPageRoute } from '../routes'
 
 import FixtureGroupsPage from './fixture-groups-page'
+import { isAnyFixtureGroupOn } from './fixture-group-actions'
 
 export const fixtureGroupsPageNavItem: NavItemEntry = {
   route: fixtureGroupsPageRoute,
   icon: iconGroup,
   label: 'Fixture Groups',
   page: FixtureGroupsPage,
-  highlighted: apiState => isAnyOn(apiState.fixtureGroups ?? {}),
+  highlighted: isAnyFixtureGroupOn,
 }