+++ frontend/src/pages/fixtures/index.ts
@@ -1,14 +1,14 @@
 import { iconLight } from '../../ui/icons'
-import { isAnyOn } from '../../util/state'
 import { NavItemEntry } from '../types'
 import { fixturesPageRoute } from '../routes'
 
 import FixturesPage from './fixtures-page'
+import { isAnyFixtureOn } from './fixtures-actions'
 
 export const fixturesPageNavItem: NavItemEntry = {
   route: fixturesPageRoute,
   icon: iconLight,
   label: 'Fixtures',
   page: FixturesPage,
-  highlighted: apiState => isAnyOn(apiState.fixtures ?? {}),
+  highlighted: isAnyFixtureOn,
 }