+++ frontend/src/pages/global/global-page.tsx
@@ -3,6 +3,7 @@ import { memoInProduction } from '../../util/development'
 import { Header } from '../../ui/containers/header'
 import { DmxMasterWidget } from '../../widgets/global/dmx-master-widget'
 import { useNumberHotkey } from '../../hooks/hotkey'
+import { ControlsWidget } from '../../widgets/global/controls-widget'
 
 /**
  * Global page.
@@ -10,13 +11,14 @@ import { useNumberHotkey } from '../../hooks/hotkey'
  * Displays global controls.
  */
 const GlobalPage = memoInProduction(() => {
-  const activeHotkeyIndex = useNumberHotkey(1)
+  const activeHotkeyIndex = useNumberHotkey(2)
 
   return (
     <>
       <Header>Global Controls</Header>
       <div className={pageWithWidgets}>
         <DmxMasterWidget hotkeysActive={activeHotkeyIndex === 0} />
+        <ControlsWidget hotkeysActive={activeHotkeyIndex === 1} />
       </div>
     </>
   )