+++ frontend/src/ui/navigation/navigation.tsx
@@ -4,7 +4,7 @@ import { mainNavigationItems } from '../../pages'
 import { iconShade, primaryShade, zNavigation } from '../styles'
 import { memoInProduction } from '../../util/development'
 import { cx } from '../../util/styles'
-import { useCompleteApiState, useMasterData } from '../../hooks/api'
+import { useApiStateSelector, useMasterData } from '../../hooks/api'
 import { iconDynamicPage } from '../icons'
 import { dynamicPageRoute } from '../../pages/routes'
 
@@ -65,24 +65,24 @@ export interface NavigationProps {
 export const Navigation = memoInProduction(
   ({ showLabels = false, floating = false }: NavigationProps) => {
     const masterData = useMasterData()
-    const apiState = useCompleteApiState(['universe'])
+    const highlightedItems = useApiStateSelector(apiState =>
+      mainNavigationItems.map(item => item.highlighted?.(apiState) ?? false)
+    )
     const { dynamicPages } = masterData
 
     return (
       <div className={cx(navigation, floating && navigation_floating)}>
-        {mainNavigationItems.map(
-          ({ route, icon, label, highlighted }, index) => (
-            <NavItem
-              key={route}
-              to={route}
-              icon={icon}
-              label={label}
-              showLabel={showLabels}
-              highlighted={highlighted?.(apiState) ?? false}
-              hotkey={navigationHotkeys[index]}
-            />
-          )
-        )}
+        {mainNavigationItems.map(({ route, icon, label }, index) => (
+          <NavItem
+            key={route}
+            to={route}
+            icon={icon}
+            label={label}
+            showLabel={showLabels}
+            highlighted={highlightedItems[index]}
+            hotkey={navigationHotkeys[index]}
+          />
+        ))}
         {dynamicPages.map(({ id, icon, headline }, index) => (
           <NavItem
             key={id}